services/requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
blinker==1.9.0
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
cryptography==45.0.7
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docutils==0.22
dotenv==0.9.9
elevenlabs==2.15.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
-e git+ssh://git@github.com/Narendrakumar-Suresh/goat-lang.git@9ce8551ef71b91d105f565c4088390542353bd5a#egg=glc
goatlang==0.0.1
google-auth==2.40.3
google-genai==1.33.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
groq==0.31.1
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
id==1.5.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
langchain-core==0.3.75
langchain-ollama==0.3.7
langsmith==0.4.25
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.4.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
nltk==3.9.1
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ollama==0.5.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.0
portalocker==3.2.0
postgrest==1.1.1
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.0.0
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
qdrant-client==1.15.1
readme_renderer==44.0
realtime==2.7.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
ruff==0.12.11
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
SecretStorage==3.3.3
semantic-version==2.10.0
sentence-transformers==5.1.0
sentry-sdk==2.37.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.3
storage3==0.12.1
streamlit==1.49.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
triton==3.4.0
twine==6.1.0
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xgboost-cpu==3.0.5
zstandard==0.24.0
